Code review changes

+++ ShareOne/Controllers/MenuViewController/DepositListViewController/DepositListController.m
@@ -259,7 +259,7 @@
 // The number of rows of data
 - (int)pickerView:(UIPickerView *)pickerView numberOfRowsInComponent:(NSInteger)component
 {
-    return _suffixArr.count;
+    return (int)_suffixArr.count;
 }
 
 // The data to return for the row and component (column) that's being passed in
